Small structural refactor, fix factory imports, update Docs
