+++ functional_tests/step_helpers/utilities.py
@@ -1,0 +1,9 @@
+def str_to_bool(bool_string: str) -> bool:
+    """Takes a string argument which indicates a boolean, and returns the corresponding boolean value.
+    raises ValueError if input string is not one of the recognized boolean like values.
+    """
+    if bool_string.lower() in ("yes", "true", "t", "y", "1"):
+        return True
+    if bool_string.lower() in ("no", "false", "f", "n", "0"):
+        return False
+    raise ValueError(f"Invalid input: {bool_string}")
